Add monitoring spec and update to template v0.2.1
There is a monitoring utility available in production which monitors the
endpoints exposed by a service. This patch adds the spec for the
MobileApps service, which instructs the utility to monitor periodically
both the standard and lite version endpoints.

Also, this patch updates the service to the latest service-template-node
v0.2.1 which comprises a test suite for validating the spec and testing
the endpoints.

Note: most node module versions have changed, so you'll need to rm -rf
node_modules && npm install for the changes to take effect.

Bug: T105772
Change-Id: I24f7c702a88cf3b17398b093ec8016634f40c923