Routes: set the ETag header
This patch makes all of the routes set the ETag header for their
responses. The format of the header is:

  revID/timeuuid

where revID is the revision ID of the transformed page and timeuuid is a
time UUID repesenting the timestamp of its creation. Note that, once
Parsoid output is used throughout the service, the provided ETag can be
used.

Change-Id: Ifb046a39128d4e0238281f894b788e0398338260