ETag: Reuse RESTBase's ETag header if available
RESTBase uses the {rev_id}/{time_uuid} format for ETags, where {rev_id}
is the revision ID and {time_uuid} is the time UUID of storing the
revision info for the given revision ID. That way RESTBase can have the
same ETag across all of the renders.

This patch allows the Mobile Content Service to reuse RESTBase's ETag
header for its own renders. The ETag is sent by RESTBase in the form of
the x-restbase-etag header.

Change-Id: I67f535d0578339514dfe516f1472e70cce6f1f30