Use Parsoid for new mobile-html-section routes
Instead of using mobileview for the page content, the
mobile-html-section* routes use Parsoid content.

Parsoid HTML is split into sections manually by looking for h2-h6 tags.
The anchors of those h[2-6] tags are built using an own implementation
of anchorencode since I wasn't able to find one online.
To help this implementation I added a dependency to the phpjs npm module.
See anchorencode in
https://www.mediawiki.org/wiki/Manual:PAGENAMEE_encoding#Encodings_compared

It still uses some of the old mobileview data, but just for some
of the metadata, not the content anymore. The mobileview call
has been changed to remove the sections and section text.

These routes now have an option path parameter "revision",
since Parsoid can return a specific revision of the page content.
Note that this is only true for the data we get from Parsoid,
meaning the page content, but currently not for the metadata.
So, some of the metadata is currently always showing the latest values.
Over time we'll try get more metadata from Parsoid.

This patch is pulling the revision and the lastmodified time stamp
from the Parsoid output. The revision is currently derived from the etag,
but I think we could also get it from the html tag about attribute.

We should look to see which pieces of metadata we can pull
out of Parsoid content. We can also add more Parsoid-specific
transformations.

Redirects:
Code was added to follow redirects from both RESTBase and Parsoid.
The redirects from Parsoid have a link tag in the payload.
The redirects coming from RESTBase have a 301 (potentially 302?) status code
(https://github.com/wikimedia/restbase/pull/365).

Step back:
* Instead of full URLs to the spokenWikipedia audio files
  we only have the "File:" names of the audio files.
  Renamed spoken.urls[] to spoken.files[] to indicate this restriction.
  I hope in the future the service could resolve the urls to the audio files automatically
  through subsequent requests. Alternatively, I think it would be cool to have an
  endpoint which provides an .m3u file of the audio files. This endpoint could resolve
  the actual audio files only when they are actually used.
  Then the client would only have one file to deal with, and we could pass this to an
  external audio player Android app.
* srcset attributes in <img> tags are not provided by Parsoid yet;
  hopefully soon. (T88827)
* Section header anchor ids had to be manually created. I hope Parsoid will provide
  them in the future: https://phabricator.wikimedia.org/T116876

Bug: T108777
Bug: T116200
Change-Id: I8cd54212f8aef6af2bc241d790221a7be9b4c199