Offload client DOM transformations
Do only the transformations that don't depend on a client setting, the client width, or are dealing with click handlers.
Since the Android app has the option of accessing the MW API directly for now,
we need to keep this in sync with the app.

Also moved the functions we use to parse certain information out of the
DOM tree into its own module, parseProperty.js.

Bug: T117450
Bug: T117505
Change-Id: I9437791ba8f5953be98546f76a6d15437cb58fc2