Add route for mobile-summary
The overlap between what we need for the lead section and previews is not
as big as originally anticipated. It's better to reduce the payload size.

This patch mainly exists to slim down the lead section request and to
preserve the infobox parsing functionality. As long as the app doesn't need it
the app could also soon use the new /page/summary/ endpoints provided by
RESTBase. See https://github.com/wikimedia/restbase/pull/401

Bug: T118147
Change-Id: I0305ae280e30a815d8d0dcd8c95490a9f183c342