Return definitions for all languages on whitelisted domain
Wiktionaries contain a large number of pages devoted to other languages'
words[0], which we hadn't anticipated when designing the service/endpoint.
This patch updates the parsing logic to return the results for all
languages when the domain is valid, in order to prevent throwing errors
for all of the pages on, e.g., English Wiktionary that don't actually
deal with English words.

N.B. This patch adds a static list of language name/code key-value pairs
for Wiktionary languages generated from the API at
https://wikistats.wmflabs.org/ using a Python script; the list was created
today and the script is included so that the list can be updated as desired.

[0] eg, https://en.wiktionary.org/wiki/%D0%B8%D0%B7%D0%B1%D0%B5%D1%80%D0%B0

Bug: T124028
Bug: T124321
Change-Id: I2bd577d1782f70081055f1082a9a5795bc444ef7