Bug fix: Limit the number of redirects followed by the service
As RESTBase does not provide the facilities to follow redirects for
pages, the service does it itself. However, it used to follow them
'blindly' without checking for redirect loops nor limiting the number of
redirects done. This patch solves the immediate problem of infinite
redirect loops by limiting the number of redirects.

As a first step, the limit is hard-coded in lib/parsoid-access.js and
set to 10. Ideally, users should be able to set it in the application's
configuration file.

Change-Id: Ie40405b0422db2af557ae267d511ad6e45ed9e55