Update domino to 1.0.23
We are seeing a substantial number of TypeErrors in production due to a
bug in Domino's CSS parsing.  The errors are not user-facing because they
appear to be exclusive to requests sent to zh.wikipedia.org, but we do not
send requests to zhwiki from the app; instead, it appears they are
generated during pregeneration.

A fix was included in 1.0.22 and deployed here on 1/28 but the fix caused
more serious issues so I rolled back the version to 1.0.21.  A follow-on
fix included in 1.0.23 should resolve the remaining problems.

A test was added to domino to verify the fix, but this commit adds a test
in the service as well for good measure.

Bug: T124446
Change-Id: I6df3024311947c83ecb75f27e9d8b0864d537c17