Build request URIs with templating from swagger-router
Requests to the beta cluster were failing since we had hard-coded the
https protocol into our request URI templates and TLS isn't properly
configured on the beta cluster.

This patch updates the dev and prod configs to use proper templating
as provided in the swagger-router library.  As for the TLS issue, we'll
send requests via http and allow the server to redirect to https if
available.

Bug: T113542
Change-Id: Ia0a8d4af247ee9f0562a895ce3824db6e85b4ccc