Hide links to missing pages, Take 2
Gets a list of links with missing pages from the VisualEditor extension API
and hides them, so we don't display dead links in the app.

This is necessary because unlike the PHP parser, Parsoid does not flag
links to nonexistent pages, so we can't just search them by class and
remove them as we did with PHP parser HTML.

Note that this will only work for wikis with VisualEditor enabled.

An earlier version of this patch was deployed and then reverted because of
memory problems in production; I believe this was because the
hideDeadLinks function was poorly designed.

Bug: T119266
Change-Id: Ib1e30775c4743e2f6dc5ea3151a0dacb51cbb001