Hide links to missing pages
Gets a list of links with missing pages from the VisualEditor extension API
and hides them, so we don't display dead links in the app.

This is necessary because unlike the PHP parser, Parsoid does not flag
links to nonexistent pages, so we can't just search them by class and
remove them as we did with PHP parser HTML.

Note that this will only work for wikis with VisualEditor enabled (which
includes at least most of the large Wikipedias).

Bug: T119266
Change-Id: I5bcaa1d1ebc4a2b1bca715b12cd86dee7635a65f