Create section divs before running JS transforms
The JavaScript transforms used in the app are meant to be applied on a section-by-
section basis, and having a means of isolating sections to perform the transforms
before splitting into the final JSON structure is therefore necessary to prevent
bugs such as T65874.

This required substantial changes to the definition parsing code, since that was
very tightly coupled with the old way of parsing sections.

This patch also splits the transforms applied for every request into two methods,
one for additive and one for destructive transforms.  This is helpful because
otherwise it's too easy to unintentionally clobber markup we've just added.

Finally, lead paragraph shifting is disabled here so it can be adapted for the new
section parsing code in a patch to follow.

Change-Id: I8c049740c390070c2be83b2efcc414c531761c9b