Add check for recent deployment commits to tagging script
Checks for existence of a configured deploy repository.

If one exists, checks it for a "recent" commit (one made within the past
two hours) and warns the user if there is none, suggesting a 'git pull'
and giving a chance to bail out.

Updated error messages to provide add'l info.

Changed some 'echo' commands to 'printf'.

Bug: T131504
Change-Id: Ibd8fec0783901e46ea22d2b182acb9ff705994d7