Aggregated endpoint cleanup
Fixes up the endpoint to get yesterday's most-read pages (as the iOS app
currently does) rather than trying to get today's when today's info is
requested and failing the whole request since the data haven't been
populated on the backend yet.

Requests for the future will still fail but that can be fixed later.

Change-Id: Id5a02f76be47cc40cd4d678297cce528f836d719