Update thumbnail property in feed endpoint responses
The feed will need both small, list-item sized images and larger "hero"
images, depending on the card.  In particular, 320px images are too small
for cards with larger images on larger, higher-res phones.

Since a list of four thumb URLs per page item is excessive, I've changed
the output format here as follows:

Feed endpoints return one (1) thumbnail URL per page/item.  This thumbnail
is requested from the MW API at the size of the larger of the thumbnail
sizes used by the two mobile apps for the type of card in question: 640px
for an article feature card, vs. 320px for an article list card.  The
larger size is requested since it's safe to rewrite a thumb URL downward in
size but not necessarily upward.  Clients may accept the thumb URL as
provided or rewrite the size downward as they see fit.

I've also introduced symbols to refer to various image widths in mwapi.js.

Bug: T129079
Change-Id: I4ae8e6bbaeb221872b1bc1d07be92f33002ff606