Most-read: Catch and handle Pageview API 404s
There's no guarantee that pageview data will be loaded for the pageview
API at any given time, and it will return a 404 if the data aren't there
yet.

This will allow us to call it safely via the aggregated endpoint and omit
the most-read section from the response in the case of a Pageview API 404.

Change-Id: Ifdf265e0aa95804df511874148ff669fc65ef5de