Update Picture of the Day endpoint
Testing revealed that this endpoint was failing aggregated requests for
languages where the POTD was unsupported under the current design.  This
is doubly flawed: (1) the aggregated request shouldn't fail in the case of
any individual component failing, but instead simply omit the failing
component; and (2) there should be no unsupported languages for this
endpoint.

Further, the unit tests were failing to catch the unsupported languages
bug because checkForQueryPagesInResponse was improperly passing through a
200 status where query.pages was missing, rather than an error code.

Further, the aggregated response was providing the previous day's featured
picture rather than the current day's, because dateUtil.yesterday was
flawed in design in that it permanently set the request object's day param
back by one, causing subsequent requests to use the previous day's date.

So, this patch accomplishes the following:

1) Stops failing the POTD request just because a caption isn't provided in
the language for the request site, and instead return the English caption
(which the client can choose to use or discard).

2) Throws a 504 if checkForQueryPagesInReponse fails rather than simply
passing through the response code.

3) Updates the POTD tests to ensure that requests that should fail are not
passing with 200s (as was happening before).

4) Removes redundant checks for query.pages after
checkForQueryPagesInResponse has already been called and ensure that
dontThrow functions correctly.

5) Removes dateUtil.yesterday and adds logic to the most-read endpoint
to make the request for the correct date.

Bug: T132765
Change-Id: Iac1d48ffaf381947f8a662585a3cdba9a70b3fa8