Ensure that tests for errors enforce an error response
Prior to Iac1d48ffaf381947f8a662585a3cdba9a70b3fa8, there were tests
that should have been failing but passed because they were receiving
a 200 response and the tests did not contain a check to ensure that
the response was in fact an error.  This adds such checks.

Change-Id: Ib776e5593cd6aa6fd108a672cc81bf06317940e6
Depends-On: Iac1d48ffaf381947f8a662585a3cdba9a70b3fa8