Add mw-format-encoded title to random & featured article endpoints
We need to include the raw, non-normalized page title in the random
endpoint response(s) in order to avoid excessive redirects when getting
the full page metadata from RESTBase.  But MediaWiki only gives us a list
of normalized titles when we request a list of random pages.

This patch leverages the mediawiki-title library to derive the raw, DB-
encoded title format based on the request site's siteinfo and includes it
in the random endpoint response.

It's also added to the featured article endpoint, which uses the same
mwapi.buildSummaryResponse function to build its output.

See https://github.com/wikimedia/restbase/pull/636#issuecomment-231385240.

Bug: T132597
Change-Id: I5eb24688a774316481b98a494d3984044d09d948