Perform Wiktionary-specific transform queries only for Wiktionary requests
Profiling reveals that stripping various content from the DOM is a long-
running and expensive set of operations in every mobile-sections* request,
and we should look for ways to streamline where possible.

This patch removes a set of selectors specific to Wiktionary from the list
of selectors used for all mobile-sections* requests and moves them to a
function called only from the Wiktionary endpoint code.

I looked here to determine which transforms were added specifically for
Wiktionary:

https://gerrit.wikimedia.org/r/#/c/255263/23/lib/transforms.js

Change-Id: I0f383d60fbaf12edd9a2aad85200d358ac4e8732