Wiktionary: Stop restricting language names to those on a static list
Wiktionary allows users to define languages arbitrarily rather than
restricting them to a fixed set.  Previously we were discarding these
where they weren't on a fixed list derived from the outstanding Wiki
languages.

This patch stops discarding "unknown" language content and instead
includes it in the response.  It does this in a strictly additive way
without changing the response structure so as to avoid breaking existing
clients.  The language code "other" is used for languages not on the
static list and the language name is included with each definition so that
the client can use it if desired.

I've also added just "Chinese" (as opposed to "Traditional Chinese" or
"Simplified Chinese") to the static list since it was absent before.

These changes should eliminate most of the remaining 404s in the
mobileapps production logs.

Bug: T129229
Bug: T129235
Change-Id: Ic07f729337599ed4038a3930d27b832c70a16a0f