Performance: Use single selector for querySelectorAll calls
Profiling reveals that our content stripping transforms are a significant
bottleneck in our processing pipeline.

It turns out that calling querySelectorAll with multiple selectors at a
time is quite slow, and performance can be improved by calling it with
a single selector even at the expense of adding additional calls.

Local testing shows this change to improve the combined time for
runAllSectionTransforms and applyOptionalParsoidSpecificTransforms by
about 148ms (18.34%).[1]

[1] Full results at https://phabricator.wikimedia.org/P3460

Change-Id: Id15dbc21b5a938b0aee745824fef0861dc734b79