spec.yaml: Disable /feed/featured tests in production
`/feed/featured` takes a long time to respond to requests, as it
aggregates a lot of sub-requests together. When scb100[12] nodes are
under stress, it takes more than 10 seconds for a test run to complete,
which creates alerts. In an effort to reduce the overall test time,
`/feed/featured` will not be monitored for the time being, since other
`featured` endpoints are monitored and return more quickly that the
aggregate endpoint.

Also, do not monitor `mobile-sections-remaining` because its result is
already contained in `mobile-sections`. Same goes for the Main_Page test
for `mobile-sections-lead`. Tests for Barack Obama and San Francisco
have been replaced with one for Altrincham as it contains both
pronunciation and spoken files as well as geo coordinates.

Additionally, disable monitoring for `mobile-text` since it's not being
actively developed.

Bug: T140215
Change-Id: I6c3d1b7f0cd53d5aa0e27202d9b2916d3cedb082