Fix date provided with most-read content in aggregated endpoint
Expected behavior for requests directly to the most-read endpoint is
to get results for the requested date, but for the aggregated endpoint
to include the previous day's top-read pages.  This is working correctly,
but due to a bug, the most-read section of the aggregated feed results
were attributing the most-read results to the aggregated endpoint request
date rather than the previous day.  This ensures that the date provided
with the most-read results is accurate in all cases.

Change-Id: If9dca73fb745effdf6b422b29fdf4670a64b0c54