Enable In the News for a few more languages.
This enables the News endpoint for additional languages of German,
Russian, and Hebrew, since these Wikipedias use a very similar template
for presenting News content on their Main Pages. So, why not?

This also tweaks the selector that gathers the anchor tags from the text,
so that it only looks for WikiLinks. Certain news stories, particularly
in Hebrew Wikipedia, include external links that should not be broken
out into our response.

Change-Id: Ie365300fd5290711ce2664a57e2d283213219c7d