Handle fragments in news links
Discard the fragment portion (i.e, the part after '#', if present) of a
page title before adding it to the list for metadata fetching.

Add title to the list only if not a duplicate.

Prevents a bug in which titles can appear twice in the list of news links
and/or without metadata.

Bug: T141307
Change-Id: Ide60041c06fdbf1171059e032dd5a2b2f56481da