Feed endpoints: Allow the client to request aggregated parts
The individual feed endpoints do not throw errors if a request has been
made to the aggregated endpoint. Since RESTBase will be using the
individual endpoints by sending them requests in parallel and then
assembling the response, enable it to signal them that the response
should be in the same format as when the aggregated endpoint is
contacted by allowing the `aggregated` query flag to be passed in.

Note that, in production, this will be available only to RESTBase, if an
external client passes it in, RESTBase will remove it from the request
when forwarding it to MCS.

Bug: T143912
Change-Id: I3ba040853e36ca53ddbf5590afd0b4430cda5749