Hatnotes should be promoted outside lead section html
The repositioning of the first paragraph in apps sends hatnotes
below that first paragraph. This change allows consumers of the API
to hide it with css and re-render it elsewhere. For example in MobileFrontend
hatnotes are rendered at the top of the page.

The hatnotes html is currently duplicated in the lead section to allow
consumers to update to use the new version of the API. This will be removed
in a follow up.

Change-Id: Iba848fd61b4bd7751934b640b70d03c089799e3d