Remove hatnotes from lead section content
When MOBILE_CONTENT_SERVICE_EDGE_VERSION environment variable is
set expose an experimental endpoint
`/formatted/:title/:revision` that mimics the existing
`/mobile-sections' endpoint but makes breaking changes.
The name of the endpoint and shape can be decided before rolling out
to production.

To kick things off the new endpoint removes hatnotes from the lead
text. The expectation is that going forward `lead.hatnotes`
is to be used to obtain this information.

Change-Id: If1dc94ec7b67e1904e6ea0e4ae025c9b612c2b9e