Add class to IPA container, do not use inline style
This is error prone and there are certain articles where this
will display incorrectly. Using a class allows consumers of the
API to handle this edge case as they wish.

This is applied inside the new /formatted endpoint but not
inside the old mobile content service for backwards compatibility

Additional changes:
* Any instance of removeNodes has been renamed and repurposed
as a legacy parameter.

Bug: T141835
Change-Id: Ica509ac9d2b643be044f4b0be3181a9bc7bc911d