Extract introduction from lead section in new experimental endpoint
After detecting the most suitable first paragraph and any associated
nodes, move it out of the lead section and into its own property
`intro`.

This will allow a consumer of the API to control positioning of it.
Along with infobox, hatnotes and the intro a consumer can
achieve all sorts of arrangements.

Bug: T145034
Change-Id: I339b398e67961cda29d12707568d9586916393ae