Fix most read test date calculation
Manipulating a single Date object is buggy and can lead to unexpected
behavior, e.g.: subtracting 5 from the date (day) part within the first
five days of the month would go to the previous month as expected, but
then setting the date forward five would result in it remaining on the
previous month.

It's likely possible to reset all subparts of the date so that these bugs
are avoided but probably easier and less error-prone to just use new
Date objects for each date wanted.

Change-Id: Ib16f84ccb29c627d0fd24b14f85bc0878a8a86b3