Do not relocate first paragraph in new endpoint
This moves the logic for relocating a lead paragraph into the routes
that serve it. Previously we did this for all endpoints but it doesn't
seem necessary for the media endpoint.

This means the transformation is not invoked in the new endpoint.

The method is also updated to take the lead as an input to avoid an
addition node lookup.

Change-Id: Iedd38dca724d62e7b9a86cf04d9858b213aae27c