Add more languages to news feed
da, el, fi, fr, ko, no, pl, sv, vi

For some languages the code needed to be changed to be more
flexible in terms of selectors used to find the correct list of
news items. (fr, da, pl)

To test I added the following paths to my local RB instance config.yaml

    /{domain:ca.wikipedia.org}: *default_project
    /{domain:el.wikipedia.org}: *default_project
    /{domain:fi.wikipedia.org}: *default_project
    /{domain:ko.wikipedia.org}: *default_project
    /{domain:no.wikipedia.org}: *default_project
    /{domain:pl.wikipedia.org}: *default_project
    /{domain:vi.wikipedia.org}: *default_project

and in the Android app pointed to that local RB instance. After switching to
a new language in the app I ended up having to refresh the feed a couple of
times before the news items showed up.

Bug: T143808
Change-Id: Iac6c156a242b1cb545ed671338bf525a07e105b4