Update news route tests
In a recent change[1], we moved to providing $merge values rather than an
article title for values to be hydrated with page summaries by RESTBase.
The value provided is the URL for the page summary, generated based on the
service config using service-runner's Template class.

This broke two tests for the news endpoint, which were therefore
temporarily marked skipped since fixing them would require
significant work.

This patch updates /lib/news.js so that the new way of handling $merge
links in the news route could be exposed in a testable way.  The tests
are updated and un-skipped once again.

[1] https://phabricator.wikimedia.org/rGMOAa359e3d3922bd3c992a8ae82267ccf0152ec0512

Change-Id: I16a44b1c03db7b32aff41624a1f7804090e35a72