Quote the ETag being set and fix it up in some endpoints
Some endpoints, like the featured image, featured page, most read and
news were emitting an ETag containing only the revision number. This
commit adds the time UUID portion needed to distinguish different
renderings of the same revision content. Additionally, the ETag needs to
be quoted as to ensure it's being interpreted as a string.

Bug: T150886
Bug: T150969
Change-Id: I4fbf8b0a7639f298213b1fbe475c522cf6e2aa19