Add additional information to file and user pages for rendering
To be able to render an image on a file page and information about
a user on the user page as in MobileFrontend the following API
adjustment is needed.

Any file pages will contain a fileinfo property.
Any user pages will contain a userinfo property.

Other properties about the image such as metadata may be useful
and we can easily add these later if necessary

Bug: T148118
Change-Id: Icf84a3b8d8ff7277ba0703a1360dd3802bc690ea