Update preq dependency
Now we can set PREQ_CONNECT_TIMEOUT when prepping the
deploy to avoid the timeout issues we've encountered
when running the tests in our slow VMs.

See https://github.com/wikimedia/preq/pull/24

The only changes since 0.5.0
https://github.com/wikimedia/preq/compare/v0.5.0...master
are the new timeout override and a travis change,
which should not affect us as a user of this library.

Change-Id: Ia22e7553d3292a5e8eb8df062818f6daacc83ffe