Make page internal links start with #
The Android app detects reference links by checking for hrefs starting with
'#'. This does not work now that Parsoid reference links start with the
current page title before the #.
This patch tries to preserve some backwards compatibility until the app
moves to the next generation of mobile-sections endpoint.
The new generation endpoint (currently called /page/formatted) is not
affected by this. So, when the Android app would move to the new
endpoint it would have to handle page internal links differently.

Bonus:
This patch started out as a fix for only the reference links but it turns
out that with a more generic selector it can also fix other page internal
links for the Android app.
A good example is on [[en:Ion]] find the first links to cation or anion,
which are links to another section on the same page.
This does not work when action=mobileview is used by the app since
there it also starts page internal links with the page title.

This patch gets the title string from the Parsoid payload then uses that
title to build a CSS selector to find all page internal links.
Some references are in a span with class="mw-ref",
while others I noticed are in a span with class="mw-reference-text".
In the former case the href starts with the title followed by '#', in the
latter case it starts with './' then the title followed by '#'.
Based on the comments in the Phabricator task it sounds like Parsoid
is going to add './' for all cases. Again, by keeping the selector less
specific it also fixes links to other sections on the same page.

Looks like domino doesn't like single or double quotes inside the selector.
I filed https://github.com/fgnass/domino/issues/95 for this.
In the meantime we work around this by cutting off the title to be used
for the CSS selector at the first occurrence of a single or double quote.
This means that we iterate over more links than necessary when on of
those quotes appear in the title string. It should be the minority of cases,
though.

Another thing that was needed to make more references work is to
stop removing the id attributes from the cite elements since some
references point to the cite element by id.

Bug: T155070
Change-Id: I7acf9d46b3867cbd1017deef39c3616990f31196