New: diff tests
Useful for regression testing. These tests do diffs of JSON output of
MCS, so we can see when things change.

The main value is in testing the /page/mobile-sections* and
/page/formatted endpoints since the bulk of transformations doesn't
receive any end-to-end testing.

To update the expected test results temporarily set the constant
UPDATE_EXPECTED_RESULTS in test-spec.js to true.
After the run consider updating the TestSpec constructors at the end of
test-spec.js to paste in the revisions and time ids (if any were left
out or need to be updated).

Also added a facility to see diffs of HTML changes easier. To enable set
ENABLE_HTML_DEBUG to true.

Bug: T155698
Change-Id: Id4706e0d13b0b4b3fb95e44bc250b8fa5a4c1748