Fix news links
anchor.href is the resolved URI of the href. It could change
depending on the value of html/head/base element. Parsoid emits it
like this: <base href="//en.wikipedia.org/wiki/"/>. So <a href="./foo">
results to '/wiki/foo'.

We probably want to use anchor.getAttribute('href') instead to have
something more predictable.

Moved removeFragment() and extracted removeLinkPrefix() to mobile-util.
I also created a convenience function for this since this is a pattern
that might be useful in other places.

This patch changes two occurrences in the news endpoint: one for
topicTitle and another for storyTitle.
I also checked for other uses of .href in the code. The only one not
changed in this patch is in mobile-text. That one seems to be actually
used correctly.

Updated news test to try to catch this if we were to run into the
same issue. There was also an issue with the if check that prevented
the asserts inside the if block to be executed.

Bug: T157040
Change-Id: I8eb2f7d8763c00d0bb1e864fc0d308315a742570