mobile-sections et al: allow Parsoid tid to be passed in
For test stability it is desirable to also specify the Parsoid time
id when making requests to mobile-sections. That way template updates
should not cause any differences in the mobile-sections output.

Added a few tests for this. Added missing headers tests.
Changed some tests in mobile-sections-v2 from formatted-lead to
formatted since that is the one which RB would use.

Change-Id: I478e6b9e91d6f4bcbc2faec5b75414e03b507448