On this day: wrap individual endpoint responses in object
By wrapping the array response in an object with the same key as used
in the 'all' response we can later use the same Swagger spec schema for
every type of endpoint. The responses would be also more consistent
between all onthisday endpoints. I just wished we would have done the
same for the other endpoints used to comprise the aggregated feed.

Note: that this schema change would require a corresponding change to
be deployed in RB at the same time.

Change-Id: I69b39501af3604da178e41521efded23302f0634