Spike: Which language projects support FeaturedFeed extension
This script uses FeaturedFeeds extension pages to get various
featured items: tfa, dyk, potd, motd, qotd, wotd, fwotd.

For tfa it gets the title of the featured article. For others it gets
the full HTML snippet.
It checks all languages to see for which languages we could get one of
these added to the aggregated feed.

The page titles are taken from
https://github.com/wikimedia/operations-mediawiki-config/blob/master/wmf-config/FeaturedFeedsWMF.php#L47-L73
(must be in MediaWiki namespace)

There are actually three modes in this script. To change the mode you
currently need to modify the script.
* wikitext of the setup page (e.g. MediaWiki:Ffeed-wotd-page)
* expanded templates
  actually expanded twice: e.g. '{{ {{ MediaWiki:Ffeed-wotd-page }} }}'
* retrieving the RSS feed using action=featuredfeed

Bug: T148680
Change-Id: I4fbb160de1be7204454cb252c1531ba03d9201fe