Fix pronunciation parsing issues
Some pronunciation file names with special characters caused issues
by creating invalid CSS selectors. In the example of
http://localhost:6927/ur.wikipedia.org/v1/page/mobile-sections/%D8%A2%D8%A6%DB%8C%D9%88%D8%B1%DB%8C_%DA%A9%D9%88%D8%B3%D9%B9
it was a single quote character but there could be others. A list can be
found in the function Sanitizer.cleanUrl
in https://github.com/wikimedia/parsoid/blob/master/lib/wt2html/tt/Sanitizer.js#L630
and phpURLEncode() in
https://github.com/wikimedia/parsoid/blob/master/lib/utils/Util.js#L1004-L1012

This is a follow up from the fix of T156941.

Also bump up the mobile-sections version for this bug fix.

Bug: T156941
Bug: T159252
Change-Id: I6fa159562c9bab1ebf24f35c4ee9e57fd5f4cd25