TFA: more languages
Expand the featured article feed entry from one language to 14.
Those are currently all languages (Wikipedia projects) which adhere
to the convention set forth by the FeaturedFeed extension, minus zhwiki
since RESTBase doesn't support language variants yet, and minus tawiki
and tgwiki, which don't really have any featured articles appearing
in the feed.

Feed/featured.js was basically rewritten to use a new approach. Now
it's using the Special:FeedItem pages provided by the FeaturedFeed
extension. It uses the query parameter useskin=apioutput as a way to
get just the content without the header and footer. Once the
FeaturedFeed extension has a real API MCS should switch to using that.

Note that for some languages there is only one TFA per week instead
of per day. Historic TFA titles on most wikis can be retrieved as far
back to the point where the FeaturedFeed extension convention was put
in place. One exception is dewiki which can only go back for the last
three days. This is a precautionary measure coming from the fact that
the template there uses {{LOCALDAYNAME}} (days of the week) to address
individual days for TFA. [1][2]

The language is usually the same as the subdomain name. One exception
of that is for Norwegian wiki we should use a different language ('nb')
than in the domain name ('no'). I think it wold work with a static
language string in most cases but better to use what the
FeaturedFeeds extension uses and be nice to the caches.

[1] https://phabricator.wikimedia.org/P5015
[2] https://de.wikipedia.org/wiki/MediaWiki:Ffeed-featured-page

Bug: T150806
Change-Id: I86eaf4634a2155d2fca3b61903d95a374e14179e