Hygiene: update diff test pages
This is a small change in strategy for the diff tests. Instead of
relying on actual pages it seems better to deal with a few
handcrafted pages for testing, which can limit the number of
templates used to just a few simple and frequently used ones which
hopefully don't change as often.

+ fix minor typo

Change-Id: If2a8eac1b4a6fc9abc475d0ddc4da844a9689c0b