Remove a few more attributes in code and img
As a side effect this also make the diff test more stable. The code
.about values were unstable. This is probably a result of adding a
new test page which includes a lot of <code> tags.

Another thing that still is unstable is the number after ImageMap.
The ImageMap is also an added feature in the new test page.
I don't want to mess with ImageMap names in production code, so I
added a post processing step to string replace the numbers after the
string ImageMap_.

Change-Id: I64258cdd96888963717eba4efcf13d75a8e2b707