Stop stripping 'rel=mw:Extlink' attributes from <a> tags
There is a transform to add class="external" to links when this appears.
It has probably been broken for some time.  When we put it back, the app
will inject an external link icon as expected.

This attribute is also required for some CSS (available in the
ResourceLoader module mediawiki.skinning.content.parsoid) to add a bit of
text with an incremented counter (e.g., '[1]') for <a> tags with external
links and empty content.  Bundling this module with the app even after
this change doesn't seem to Just Work, though.  Still, the links will at
least be clickable due to the icon now appearing.

Bug: T162116
Change-Id: I44a28b0b3c339064df7559d336d347ffdb84731d