Wrap video elements in anchor elements
Look for video elements, and wrap them inside an anchor with the right
class and the href set to the File title of the resource so that it
opens up the video in the Android app gallery.

It's a bit of a hack to make existing Android app versions handle the
video elements. Newer versions should handle video elements directly.

Bump mobile-sections patch version.

Bug: T164522
Change-Id: Ic45c0254706e345a3844d335aab01aa88fbc662c