Limit video transform to videos
This is a follow-up to Ic45c0254.

Parsoid uses video tags for both video and audio files. The previous
logic did not distinguish between video and audio files and treated
them the same. This was incorrect since audio files are to be played
embedded while videos go to the Gallery.

Bug: T167183
Bug: T164522
Change-Id: Iac99237d97ccdc76141d9f9ae0b783dfcd76adf6