Add scripts to measure payloads
We want to compare the Parsoid HTML with the read-html endpoint. For the
latter we can also vary between turning off the payload reducing
transformations and stripping of references.

In the setup phase (-s) the PageView API is consulted to retrieve the top
monthly page titles, then a request to every page is made to retrieve the
revision and time uuid. Once they are all collected a json file is written
with this information.

When the -s commandline switch is omitted the measuring phase is started.
It reads the page titles and revisions/time uuids for each entry and
requests the parsoid and the local MCS payloads.

The resulting output can be imported into a spreadsheet for further analysis.
The separator between the values is a tab. (I first tried commas
but some title include commas so that was problematic.)

Bug: T164033
Change-Id: I0466c664ea485fed5fb0bf176fdeb39f5426df78