Fix pronunciation parsing
T169293 has been fixed on the Parsoid side by bringing back the links
to the audio files. They are now in a different portion of the Parsoid
HTML now, which allows MCS to simplify pronunciation parsing code.
We only need to look for a rel=mw:MediaLink anchor in next sibling of
the IPA spans.

escape-string-regexp is not needed anymore in the parseProperty module
but still use somewhere else.

Also: Brings back pronunciation tests
Revert "Temporarily disable pronunciation tests"

This reverts commit 07ef2e8be08dbdd2c9627ab762711ae1ad233d5f.

Bug: T170325
Change-Id: I5de58afe7c2d5980bab18e5dfeb73fd30fd08021