Fix broken tests (do not skip)
Skipping should only be used as a short term measure.
It's an anti-pattern - if tests are broken they should be either
fixed promptly or removed. They may be masking real bugs that
deserve debugging attention.

Since articles are living breathing documents these integration
tests should be run against the revisions they were written
against where possible.
This updates all queries to pass a known safe revision
to test against where this fixes the problem

"Page with IPA content" is not related to T170325 so is fixed.

Where a test is being skipped because of an upstream problem
mark it with the bug number.

Change-Id: I6afd85f0f82df7d40ea7f84d0fa49a4fb6f3f94e