Bring back anchorencode and locutus dependency
This is in preparation for Iebf446c2. When using the action API
instead of mobileview redirects.tofragment is not encoded.
While this may be fine eventually, once Parsoid implements T152540.
It's a bit early for now. Once Parsoid also implements T152540
the usage of anchorencode can be removed and this commit could
be reverted.

FYI, this reverts some of the changes done earlier in I4a78ead3 and 
Ib774f286.

Bug: T103362
Change-Id: I7fc075d8626c07424662265226ad7f13d582ed7d